fix(expert-skill): keep FlyDSL MoE recipes version-portable
Remove hard runtime gating so newer FlyDSL environments can reproduce the validated optimization mechanisms and remeasure them on-box.
